+++ src/types/Methods.ts
@@ -2,4 +2,5 @@ export type Methods = {
   onInit?: () => void;
   onRandomize?: () => void;
   onResult?: () => void;
+  onRangeEnd?: () => void;
 };