+++ CHANGELOG.md
@@ -3,4 +3,4 @@
 ## 1.0.0 beta 1 (Soon)
 #### New Feature
 - added settings: `min`, `max`, `exclude`,`hold`
-- added methods: `onInit`, `onRandomize`, `onResult`
+- added methods: `onInit`, `onRandomize`, `onResult`, `onRangeEnd`